pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "batchflow"
version = "0.1.0"
description = "Async DAG workflow orchestration for HTCondor/BPS batch systems"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
webhook = ["aiohttp>=3.9"]
inotify = ["watchfiles>=0.21"]
postgres = ["asyncpg>=0.29"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
    "pytest-timeout",
]

[project.scripts]
batchflow = "batchflow.cli:main"
eo-pipe   = "eo_pipe_adapter.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["batchflow*", "eo_pipe_adapter*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
timeout = 30
